cryptoMarketCap24Hour

+++ Server/src/main/java/com/example/cryptoworld/service/CryptoService.java
@@ -3,6 +3,7 @@ package com.example.cryptoworld.service;
 import com.example.cryptoworld.models.view.CryptoViewModel;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CryptoService {
@@ -10,5 +11,5 @@ public interface CryptoService {
 
     List<CryptoViewModel> getAllCrypto();
 
-    double cryptoMarketCap24Hour ();
+    BigDecimal cryptoMarketCap24Hour ();
 }
